Fix breaking changes
