+++ api/src/main/java/io/strimzi/api/kafka/model/KafkaConnect.java
@@ -81,7 +81,7 @@ import static java.util.Collections.unmodifiableList;
 @Version(Constants.V1BETA2)
 @Group(Constants.RESOURCE_GROUP_NAME)
 @SuppressFBWarnings("RCN_REDUNDANT_NULLCHECK_OF_NONNULL_VALUE")
-public class KafkaConnect extends CustomResource<KafkaConnectSpec, KafkaConnectStatus> implements Namespaced, UnknownPropertyPreserving, RackAware {
+public class KafkaConnect extends CustomResource<KafkaConnectSpec, KafkaConnectStatus> implements Namespaced, UnknownPropertyPreserving {
     private static final long serialVersionUID = 1L;
 
     public static final String SCOPE = "Namespaced";
@@ -184,11 +184,5 @@ public class KafkaConnect extends CustomResource<KafkaConnectSpec, KafkaConnectS
     @Override
     public void setAdditionalProperty(String name, Object value) {
         this.additionalProperties.put(name, value);
-    }
-
-    @Override
-    @Description("Rack awareness configuration.")
-    public Rack getRack() {
-        return spec.getRack();
     }
 }
