--- api/src/main/java/io/strimzi/api/kafka/model/RackAware.java
@@ -1,9 +1,0 @@
-/*
- * Copyright Strimzi authors.
- * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
- */
-package io.strimzi.api.kafka.model;
-
-public interface RackAware {
-    Rack getRack();
-}
